Reland another version of "The Sema::LookupConstructor is not iteration safe."
When looking up a ctor the modules infrasturcture deserializes more ctor
candidates in the body of the function causing the internal vector implementation
to rellocate and invalidate the pointers.

This patch makes Sema::LookupConstructor void and stabilizes the iteration.

This patch is a previous version of the patch in https://reviews.llvm.org/D91524
and we, after being merged, we should be able to backport it.